Fix warnings and create theme change depending on user settings

+++ SymbolSpeakAAC/app/src/main/java/com/example/symbolspeak_aac/TextToSpeach/TextToSpeechViewModel.kt
@@ -4,10 +4,8 @@ import androidx.compose.runtime.mutableStateOf
 import android.content.Context
 import android.speech.tts.TextToSpeech
 import androidx.compose.runtime.State
-import androidx.compose.runtime.collectAsState
 import androidx.lifecycle.ViewModel
-import com.example.symbolspeak_aac.Symbol
-import com.example.symbolspeak_aac.UserSettings
+import com.example.symbolspeak_aac.Symbol.Symbol
 import java.util.*
 
 class TextToSpeechViewModel : ViewModel() {
@@ -35,7 +33,7 @@ class TextToSpeechViewModel : ViewModel() {
             if (it == TextToSpeech.SUCCESS) {
                 textToSpeech?.let { txtToSpeech ->
                     txtToSpeech.language = Locale.ENGLISH
-                    txtToSpeech.setSpeechRate(ttsRate.toFloat())
+                    txtToSpeech.setSpeechRate(ttsRate)
                     txtToSpeech.speak(
                         text,
                         TextToSpeech.QUEUE_ADD,
